2016-7-8  大修整1版

+++ company/Invest/Model/InvestListModel.h
@@ -20,7 +20,7 @@
 @property (nonatomic, copy) NSString *introduce;
 
 @property (nonatomic, strong) NSArray<NSString *> *areas;
-@property (nonatomic, assign) NSInteger collectCount; // 关注数量
+@property (nonatomic, assign) int collectCount; // 关注数量
 @property (nonatomic, assign) BOOL collected; //是否关注
 
 @property (nonatomic, assign) BOOL commited;
